third_party/blink/web_tests/wpt_internal/webid/get-argument-validation.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>WebID navigator.id.get() inputs Tests</title>
<link rel="author" title="Majid Valipour" href="mailto:majidvp@chromium.org">
<link rel="help" href="Test basic parameter validation for navigator.id.get()">
<link rel="help" href="https://wicg.github.io/WebID">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
  // request
  promise_test(async t => {
    const result = navigator.id.get({
      provider: 'https://idp.test',
    });
    return promise_rejects_js(t, TypeError, result);
  }, "Reject when request is missing");

  // provider
  promise_test(async t => {
    const result = navigator.id.get({
      request: ''
    });
    return promise_rejects_js(t, TypeError, result);
  }, "Reject when provider is missing" );

  promise_test(async t => {
    const result = navigator.id.get({
      provider: 'not_a_url',
      request: '',
    });
    return promise_rejects_dom(t, 'SyntaxError', result);
  }, `Reject when provider is an invalid URL`);

  // mode
  promise_test(async t => {
    const result = navigator.id.get({
      provider: 'https://idp.test',
      request: '',
      mode: 'invalid_mode'
    });
    return promise_rejects_js(t, TypeError, result);
  }, "Reject when mode is invalid");
</script>
